Split up webServer.ts
Refactors webServer.ts to split it into multiple files and encapsulate some functionality in classes
