+++ extensions/typescript-language-features/web/src/typingsInstaller/jsTyping.ts
